Move web_view_manager to atom/browser
