+++ atom.gyp
@@ -190,8 +190,8 @@
       'atom/browser/ui/x/window_state_watcher.h',
       'atom/browser/ui/x/x_window_utils.cc',
       'atom/browser/ui/x/x_window_utils.h',
-      'atom/browser/web_view/web_view_manager.cc',
-      'atom/browser/web_view/web_view_manager.h',
+      'atom/browser/web_view_manager.cc',
+      'atom/browser/web_view_manager.h',
       'atom/browser/web_dialog_helper.cc',
       'atom/browser/web_dialog_helper.h',
       'atom/browser/window_list.cc',
