+++ atom/browser/web_view_manager.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_WEB_VIEW_WEB_VIEW_MANAGER_H_
-#define ATOM_BROWSER_WEB_VIEW_WEB_VIEW_MANAGER_H_
+#ifndef ATOM_BROWSER_WEB_VIEW_MANAGER_H_
+#define ATOM_BROWSER_WEB_VIEW_MANAGER_H_
 
 #include <map>
 
@@ -100,4 +100,4 @@ class WebViewManager : public content::BrowserPluginGuestManager {
 
 }  // namespace atom
 
-#endif  // ATOM_BROWSER_WEB_VIEW_WEB_VIEW_MANAGER_H_
+#endif  // ATOM_BROWSER_WEB_VIEW_MANAGER_H_
